+++ sdk/containerapps/Azure.Provisioning.AppContainers/tests/BasicAppContainersTests.cs
@@ -4,7 +4,6 @@
 using System.Threading.Tasks;
 using Azure.Provisioning.Expressions;
 using Azure.Provisioning.OperationalInsights;
-using Azure.Provisioning.Primitives;
 using Azure.Provisioning.Tests;
 using NUnit.Framework;
 
@@ -12,25 +11,6 @@ namespace Azure.Provisioning.AppContainers.Tests;
 
 public class BasicAppContainersTests
 {
-    [Test]
-    public void ManagedEnvironmentNameRequirements()
-    {
-        ResourceNameRequirements requirements =
-            new ContainerAppManagedEnvironment("env").GetResourceNameRequirements();
-
-        Assert.Multiple(() =>
-        {
-            Assert.That(requirements.MinLength, Is.EqualTo(2));
-            Assert.That(requirements.MaxLength, Is.EqualTo(60));
-            Assert.That(
-                requirements.ValidCharacters,
-                Is.EqualTo(
-                    ResourceNameCharacters.LowercaseLetters |
-                    ResourceNameCharacters.Numbers |
-                    ResourceNameCharacters.Hyphen));
-        });
-    }
-
     internal static Trycep CreateContainerAppTest()
     {
         return new Trycep().Define(