+++ sdk/containerapps/Azure.Provisioning.AppContainers/CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Bugs Fixed
 
-- Corrected generated names for Container Apps managed environments to use 2–60 lowercase letters, numbers, or hyphens.
+- Corrected default names for Container Apps managed environments to use 2–60 lowercase letters, numbers, or hyphens.
 
 ### Other Changes